add javascript extensions to the mix for extending applications
JavaScript Extensions allow for applications which include the
myfedora.extensions.js file to their page to place extention points within
the web page which will get populated by extensions which consume specific
types of extension points.  There are a number of parts to this framework.

1) The Loader - Not Implemented Yet

The loader loads all extensions and catagorises them based on the types they
consume.

2) The Controller

The controller is mounted on extensions/  and serves up each of the extensions
based on the exttype passed to it.  To avoid roundtrips it will return all
extensions for a given type in a list.

3) The Extension

Extensions are pure JavaScript and must be implemented in a single hash "class"
with a run function as one of its memebers.  The run function must take one
parameter which is the data passed to it via the extension point and return
the HTML it wishes to be grafted onto the extension point.  JQuery and
myfedora javascript modules are garuntteed to be available.  Other libraries
must be included in the app which is being extended in order to be used.

4) The Extension Point

Extension points are script tags with the text/x-myfedora-extpoint type.  They
should contain valid json data wich will be passed to the extension's run
method.  Data that is manditory are as follows:

* type - the type of extensions which can be displayed
* placeholder_id - the id of the block which the extension will be appending
                   its output to

Included with the manditory data is also explicit data that is filled in by the
extension engine:

* uid - the unique id of the block where the extension data will be placed
        which is generated before running the extension.  While the placeholder
        id tells the engine where to place the HTML output the uid allows the
        extension to continuiously update its own content without effecting
        other extensions being placed on the same placeholder.
        This is available before the extension is run so the extension should
        be able to grab the div and style it.