Simplify dogpile cache interfaces.
There used to be two unique ones.  One for all the connectors, and a second one
just for the bugzilla connector.  This was a holdover from when I was trying to
optimize bugzilla queries by caching them bit-by-bit as they came in.  This
isn't necessary anymore since the ssl timeout error has been fixed.