Use python-retask to distribute cache refreshing to a worker proc.
This is a first revision/proof-of-concept and needs a lot of cleanup.
It works though!  Async refills of the cache are marked in a redis queue.
A worker proc picks tasks out of that queue, computes the value, places
the new value in the memcached cache, and releases the distributed
dogpile mutex.