Fix bugzilla + tw2.
An update to tw2.core upstream was causing problems with the 'bugs' tab
in development.  Basically, the semantics of tw2.core.Param() became
more restrictive, meaning that the package name would have to be passed
in as an explicit keyword argument.  fedora-packages doesn't do that, so
the BugsWidget was failing with a traceback from tw2's validation code.
However, the BugsWidget was the *only* package widget from
fedora-packages that had this Param declared.. so I'm just removing it
to put it in line with all the other widgets.. and to get everything
working again.