Fix tw2 resource path when mounted outside siteroot
When the application is mounted outside of the site root (/), we need to pass
TW2 its relative URL to the root.
For TurboGears before 2.1, this is done by add_tosca2_middleware, but since
TG2.1, the initialization of TW2 is performed by TurboGears itself.
This injects the most important configuration option from the global config
into tw2 configuration.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>