problems: Fix filtering of packages in Problems tab
The filtering didn't filter by package_name. It queried all problems for all packages from FAF
which resulted in timeouts.

With this commit, the filtering should use the name of package that is being currently viewed and
use it to query problems from FAF.

This should reduce:
a) query time from FAF
b) size of returned data from FAF

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>